Moves QiskitElementPage and QiskitPage  to logic folder
