+++ components/logic/QiskitElementPage.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
-import QiskitPage from '~/components/qiskit/QiskitPage.vue'
+import QiskitPage from '~/components/logic/QiskitPage.vue'
 
 @Component
 export default abstract class extends QiskitPage {
