+++ pages/events/_slug.vue
@@ -33,7 +33,7 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import { Context } from '@nuxt/types'
-import QiskitPage from '~/components/qiskit/QiskitPage.vue'
+import QiskitPage from '~/components/logic/QiskitPage.vue'
 import EventMenu from '~/components/events/EventMenu.vue'
 import EventFooter from '~/components/events/EventFooter.vue'
