+++ pages/events/index.vue
@@ -99,7 +99,7 @@
 <script lang="ts">
 import { mapGetters, mapActions } from 'vuex'
 import { Component } from 'vue-property-decorator'
-import QiskitPage from '~/components/qiskit/QiskitPage.vue'
+import QiskitPage from '~/components/logic/QiskitPage.vue'
 import EventCard from '~/components/events/EventCard.vue'
 import {
   CommunityEvent,
