+++ pages/experiments/_experiment.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import { Context } from '@nuxt/types'
-import QiskitPage from '~/components/qiskit/QiskitPage.vue'
+import QiskitPage from '~/components/logic/QiskitPage.vue'
 import ExperimentHeader from '~/components/experiments/ExperimentHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
 import MdContent from '~/components/ui/MdContent.vue'
