+++ pages/ibmqaccount.vue
@@ -56,7 +56,7 @@ print(backend, backend.status().pending_jobs)</SyntaxHighlight>
 
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
-import QiskitElementPage from '~/components/elements/QiskitElementPage.vue'
+import QiskitElementPage from '~/components/logic/QiskitElementPage.vue'
 import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
 import LegacySection from '~/components/sections/LegacySection.vue'
 import SoftwareStack from '~/components/ui/SoftwareStack.vue'
