+++ pages/ignis.vue
@@ -91,7 +91,7 @@ for rb_seed,rb_circ_seed in enumerate(rb_circs):
 
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
-import QiskitElementPage from '~/components/elements/QiskitElementPage.vue'
+import QiskitElementPage from '~/components/logic/QiskitElementPage.vue'
 import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
 import LegacySection from '~/components/sections/LegacySection.vue'
 import SoftwareStack from '~/components/ui/SoftwareStack.vue'
