+++ pages/terra.vue
@@ -70,7 +70,7 @@ print(sim_result.get_counts(qc))</SyntaxHighlight>
 
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
-import QiskitElementPage from '~/components/elements/QiskitElementPage.vue'
+import QiskitElementPage from '~/components/logic/QiskitElementPage.vue'
 import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
 import LegacySection from '~/components/sections/LegacySection.vue'
 import SoftwareStack from '~/components/ui/SoftwareStack.vue'
